+++ src/main/java/org/gridsuite/shortcircuit/server/service/ShortCircuitReportMapper.java
@@ -24,7 +24,7 @@ import java.util.regex.Pattern;
  * This class manages how to postprocess reports of the proprietary courcir simulator to reduce the number of reports
  * by aggregating them.
  * <br/>
- * The tree structure created by the courcirc proprietary simulator is:
+ * The tree structure returned by the courcirc proprietary simulator is:
  * <pre>
  * 00000000-0000-0000-0000-000000000000@ShortCircuitAnalysis
  * \-- ShortCircuitAnalysis -> "ShortCircuitAnalysis (${providerToUse})"
@@ -36,7 +36,7 @@ import java.util.regex.Pattern;
  *     |   +-- twoWindingsTransformerConversion -> "Conversion of two windings transformers"
  *     |   \-- threeWindingsTransformerConversion -> "Conversion of three windings transformers"
  *     +-- danglinglinesConversion -> "Conversion of the dangling lines"
- *     \-- extkey -> "Logs generated by an extension (provider)"
+ *     \--courcirc -> "Logs generated by Courcirc simulator"
  * </pre>
  *
  * @implNote as the tree to aggregate seems relatively simple, a simple utility class will suffice
@@ -45,7 +45,6 @@ import java.util.regex.Pattern;
  * @see com.powsybl.commons.reporter.ReporterModel
  * @see com.powsybl.commons.reporter.Report
  */
-@SuppressWarnings("SwitchStatementWithTooFewBranches") //for future additions
 @Slf4j
 @Component
 public class ShortCircuitReportMapper {
@@ -54,10 +53,10 @@ public class ShortCircuitReportMapper {
      * @param reporter the reporter to modify
      * @return the result
      *
-     * @apiNote is limited to some implementations of {@link Reporter}
+     * @implNote currently support only some implementations of {@link Reporter}
      */
     public Reporter modifyReporter(@NonNull final Reporter reporter) {
-        if (reporter instanceof ReporterModel reporterModel && reporterModel.getTaskKey().endsWith("ShortCircuitAnalysis")) {
+        if (reporter instanceof ReporterModel reporterModel && reporterModel.getTaskKey().matches("^([a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}@)?ShortCircuitAnalysis$")) {
             log.debug("ShortCircuitAnalysis root node found, will modify it!");
             return forUuidAtShortCircuitAnalysis(reporterModel);
         } else {
@@ -76,7 +75,7 @@ public class ShortCircuitReportMapper {
         reporterModel.getReports().forEach(newReporter::report);
         reporterModel.getSubReporters().forEach(reporter -> {
             if ("ShortCircuitAnalysis".equals(reporter.getTaskKey())
-                && "courcirc".equals(reporter.getTaskValues().getOrDefault("providerToUse", new TypedValue("", "")).getValue())) {
+                && "Courcirc".equals(reporter.getTaskValues().getOrDefault("providerToUse", new TypedValue("", "")).getValue())) {
                 newReporter.addSubReporter(forShortCircuitAnalysis(reporter));
             } else {
                 newReporter.addSubReporter(reporter);
@@ -104,17 +103,18 @@ public class ShortCircuitReportMapper {
     /**
      * Modify node with key {@code courcirc}
      * <br/>
-     * The relevant part of the tree structure created by the courcirc proprietary simulator is:
+     * The relevant part of the tree structure created by the courcirc proprietary simulator output is:
      * <pre>
      * 00000000-0000-0000-0000-000000000000@ShortCircuitAnalysis
      * \-- ShortCircuitAnalysis -> "ShortCircuitAnalysis (${providerToUse})"
      *     \-- courcirc -> "Logs generated by Courcirc simulator"
      * </pre>
-     * this courcirc reporter, among other usefull reports, contains hundreds of reports that are counted and replaced
+     * this courcirc reporter, among other useful reports, contains hundreds of reports that are counted and replaced
      * by a single aggregate report
      *
      * @apiNote the main goal here is to concat specific log lines with a huge number of repetitions
-     * @implNote courcirc logs don't seem to use enumerate keys (REC_0,REC_1,...) neither values, so we need to test the message
+     * @implNote courcirc logs don't seem to use enumerate keys neither template value, only incremental keys (REC_0,REC_1,...),
+     *           so we need to test the message
      */
     private ReporterModel forCourcirc(@NonNull final ReporterModel reporterModel) {
         log.trace("courcirc logs detected, will analyse them...");
