+++ src/test/java/org/gridsuite/shortcircuit/server/ReporterLogsTest.java
@@ -6,11 +6,10 @@
  */
 package org.gridsuite.shortcircuit.server;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.powsybl.commons.reporter.Report;
 import com.powsybl.commons.reporter.Reporter;
 import com.powsybl.commons.reporter.ReporterModel;
-import com.powsybl.commons.reporter.TypedValue;
 import com.powsybl.computation.ComputationManager;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.VariantManager;
@@ -19,13 +18,19 @@ import com.powsybl.network.store.client.PreloadingStrategy;
 import com.powsybl.shortcircuit.*;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.text.StringSubstitutor;
+import org.assertj.core.api.Assertions;
 import org.assertj.core.api.WithAssertions;
 import org.assertj.core.api.recursive.comparison.RecursiveComparisonConfiguration;
+import org.assertj.core.presentation.StandardRepresentation;
 import org.gridsuite.shortcircuit.server.repositories.ShortCircuitAnalysisResultRepository;
 import org.gridsuite.shortcircuit.server.service.*;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.MockedStatic;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -33,8 +38,8 @@ import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.support.GenericMessage;
 
+import java.io.IOException;
 import java.util.List;
-import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
 import java.util.stream.Stream;
@@ -52,44 +57,18 @@ class ReporterLogsTest implements WithAssertions {
     private final ShortCircuitReportMapper reportMapper = new ShortCircuitReportMapper();
 
     static final String ROOT_REPORTER_ID = "00000000-0000-0000-0000-000000000000@ShortCircuitAnalysis";
+    static final String ROOT_REPORTER_NO_ID = "ShortCircuitAnalysis";
     static final String SHORTCIRCUIT_TYPE_REPORT = "ShortCircuitAnalysis";
 
     private static ReporterModel rootReporter;
-    private static RecursiveComparisonConfiguration assertjRecursiveComparisonConfiguration = RecursiveComparisonConfiguration.builder()
+    private static final RecursiveComparisonConfiguration ASSERTJ_RECURSIVE_COMPARISON_CONFIGURATION = RecursiveComparisonConfiguration.builder()
             .withIgnoreCollectionOrder(false)
             .withIgnoreAllOverriddenEquals(true)
             .build();
 
     @BeforeAll
-    static void prepare() {
-        rootReporter = new ReporterModel(ROOT_REPORTER_ID, ROOT_REPORTER_ID);
-        final Reporter reporter = rootReporter.createSubReporter(SHORTCIRCUIT_TYPE_REPORT, SHORTCIRCUIT_TYPE_REPORT + " (${providerToUse})", "providerToUse", "Courcirc");
-        reporter.createSubReporter("generatorConversion", "Conversion of generators")
-                .report("disconnectedTerminalGenerator", "Regulating terminal of connected generator ${generator} is disconnected. Regulation is disabled.", "generator", "TestGenerator");
-        {
-            final Reporter reporter1 = reporter.createSubReporter("batteryConversion", "Conversion of the batteries");
-            reporter1.createSubReporter("lineConversion", "Conversion of lines");
-            reporter1.createSubReporter("tieLineConversion", "Conversion of tie lines");
-            reporter1.createSubReporter("twoWindingsTransformerConversion", "Conversion of two windings transformers");
-            reporter1.createSubReporter("threeWindingsTransformerConversion", "Conversion of three windings transformers");
-        }
-        reporter.createSubReporter("danglinglinesConversion", "Conversion of the dangling lines");
-        final Reporter reporterCourcirc = reporter.createSubReporter("courcirc", "Logs generated by Courcirc simulator");
-        {
-            int recNb = 0;
-            reporterCourcirc.report("REC_" + recNb++, "Taking into account resistances  : oui");
-            reporterCourcirc.report("REC_" + recNb++, "Taking into account capacities   : oui");
-            reporterCourcirc.report("REC_" + recNb++, "Taking into account VSC modelisation : non");
-            reporterCourcirc.report("REC_" + recNb++, "Taking into account current ratio of tap changers : oui");
-            reporterCourcirc.report("REC_" + recNb++, "Taking into account loads        : oui");
-            reporterCourcirc.report("REC_" + recNb++, "Taking into account shunts       : non");
-            reporterCourcirc.report("REC_" + recNb++, "Taking into account mutual lines : non");
-            reporterCourcirc.report("REC_" + recNb++, "Taking into account sub-transient values : non");
-            reporterCourcirc.report("REC_" + recNb++, "Interior area = whole network");
-            for (int j = 0; j < 5; j++) {
-                reporterCourcirc.report("REC_" + (recNb++), "T.NODE" + j + " : transient reactance too low ==> generator ignored");
-            }
-        }
+    static void prepare() throws IOException {
+        rootReporter = RestTemplateConfig.objectMapper().readValue(ReporterLogsTest.class.getClassLoader().getResource("reporter_courcirc_test.json"), ReporterModel.class);
     }
 
     @Test
@@ -103,51 +82,28 @@ class ReporterLogsTest implements WithAssertions {
         assertThat(reportMapper.modifyReporter(reporter)).isSameAs(reporter);
     }
 
-    @Test
-    void testModifyRootNode() {
-        final Reporter targetReporter = new ReporterModel(ROOT_REPORTER_ID, ROOT_REPORTER_ID);
+    @ParameterizedTest
+    @ValueSource(strings = {ROOT_REPORTER_ID, ROOT_REPORTER_NO_ID})
+    void testModifyRootNode(final String rootId) {
+        final Reporter targetReporter = new ReporterModel(rootId, rootId);
         final Reporter reporter = targetReporter.createSubReporter(SHORTCIRCUIT_TYPE_REPORT, SHORTCIRCUIT_TYPE_REPORT + " (${providerToUse})", "providerToUse", "Courcirc");
         assertThat(reporter).isInstanceOf(ReporterModel.class);
         reporter.createSubReporter("generatorConversion", "Conversion of generators")
                 .report("disconnectedTerminalGenerator", "Regulating terminal of connected generator ${generator} is disconnected. Regulation is disabled.", "generator", "TestGenerator");
         assertThat(reportMapper.modifyReporter(targetReporter))
                 .isNotSameAs(targetReporter)
-                .usingRecursiveComparison(assertjRecursiveComparisonConfiguration)
+                .usingRecursiveComparison(ASSERTJ_RECURSIVE_COMPARISON_CONFIGURATION)
                 .isEqualTo(targetReporter);
     }
 
     @Test
-    void testModificationLogs() throws JsonProcessingException {
-        final ReporterModel targetReporter = new ReporterModel(ROOT_REPORTER_ID, ROOT_REPORTER_ID);
-        final Reporter reporter = targetReporter.createSubReporter(SHORTCIRCUIT_TYPE_REPORT, SHORTCIRCUIT_TYPE_REPORT + " (${providerToUse})", "providerToUse", "Courcirc");
-        reporter.createSubReporter("generatorConversion", "Conversion of generators")
-                .report("disconnectedTerminalGenerator", "Regulating terminal of connected generator ${generator} is disconnected. Regulation is disabled.", "generator", "TestGenerator");
-        final Reporter reporter1 = reporter.createSubReporter("batteryConversion", "Conversion of the batteries");
-        reporter1.createSubReporter("lineConversion", "Conversion of lines");
-        reporter1.createSubReporter("tieLineConversion", "Conversion of tie lines");
-        reporter1.createSubReporter("twoWindingsTransformerConversion", "Conversion of two windings transformers");
-        reporter1.createSubReporter("threeWindingsTransformerConversion", "Conversion of three windings transformers");
-        reporter.createSubReporter("danglinglinesConversion", "Conversion of the dangling lines");
-        final Reporter reporterCourcirc = reporter.createSubReporter("courcirc", "Logs generated by Courcirc simulator");
-        reporterCourcirc.report("REC_0", "Taking into account resistances  : oui");
-        reporterCourcirc.report("REC_1", "Taking into account capacities   : oui");
-        reporterCourcirc.report("REC_2", "Taking into account VSC modelisation : non");
-        reporterCourcirc.report("REC_3", "Taking into account current ratio of tap changers : oui");
-        reporterCourcirc.report("REC_4", "Taking into account loads        : oui");
-        reporterCourcirc.report("REC_5", "Taking into account shunts       : non");
-        reporterCourcirc.report("REC_6", "Taking into account mutual lines : non");
-        reporterCourcirc.report("REC_7", "Taking into account sub-transient values : non");
-        reporterCourcirc.report("REC_8", "Interior area = whole network");
-        reporterCourcirc.report("TransientReactanceTooLow", "${nb} node(s) with transient reactance too low ==> generator ignored\n${nodes}",
-                Map.of("reportSeverity", TypedValue.WARN_SEVERITY,
-                       "nb", new TypedValue(5, TypedValue.UNTYPED),
-                       "nodes", new TypedValue("T.NODE0, T.NODE1, T.NODE2, T.NODE3, T.NODE4", TypedValue.UNTYPED)));
-
+    void testModificationLogs() throws IOException {
+        final ReporterModel targetReporter = RestTemplateConfig.objectMapper().readValue(ReporterLogsTest.class.getClassLoader().getResource("reporter_courcirc_modified.json"), ReporterModel.class);
         final Reporter result = reportMapper.modifyReporter(rootReporter);
         log.debug("Result = {}", Jackson2ObjectMapperBuilder.json().findModulesViaServiceLoader(true).build().writerWithDefaultPrettyPrinter().writeValueAsString(result));
         assertThat(result)
                 .isNotSameAs(rootReporter)
-                .usingRecursiveComparison(assertjRecursiveComparisonConfiguration)
+                .usingRecursiveComparison(ASSERTJ_RECURSIVE_COMPARISON_CONFIGURATION)
                 .isEqualTo(targetReporter);
     }
 
@@ -216,6 +172,31 @@ class ReporterLogsTest implements WithAssertions {
         @Override
         public CompletableFuture<ShortCircuitAnalysisResult> run(Network network, List<Fault> faults, ShortCircuitParameters parameters, ComputationManager computationManager, List<FaultParameters> faultParameters, Reporter reporter) {
             return CompletableFuture.completedFuture(this.result);
+        }
+    }
+
+    @BeforeAll
+    static void config() {
+        Assertions.useRepresentation(new ReportRepresentation());
+    }
+
+    /**
+     * A {@code toString()} implementation a little more useful than "{@code com.powsybl.commons.reporter.Report@14998e21}"
+     * in assertion messages.
+     */
+    public static class ReportRepresentation extends StandardRepresentation {
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public String toStringOf(Object object) {
+            if (object instanceof Report report) {
+                return "@" + StringUtils.rightPad(Integer.toHexString(System.identityHashCode(report)), 9)
+                        + (report.getValues().containsKey("reportSeverity") ? report.getValue("reportSeverity").getValue() : "UNKOWN")
+                        + " [" + report.getReportKey() + "] " + StringSubstitutor.replace(report.getDefaultMessage(), report.getValues());
+            } else {
+                return super.toStringOf(object);
+            }
         }
     }
 }
