+++ src/test/resources/reporter_courcirc_modified.json
@@ -1,0 +1,135 @@
+{
+  "version": "1.0",
+  "reportTree": {
+    "taskKey" : "00000000-0000-0000-0000-000000000000@ShortCircuitAnalysis",
+    "taskValues" : { },
+    "reports" : [ ],
+    "subReporters" : [ {
+      "taskKey" : "ShortCircuitAnalysis",
+      "taskValues" : {
+        "providerToUse" : {
+          "value" : "Courcirc",
+          "type" : "UNTYPED"
+        }
+      },
+      "reports" : [ ],
+      "subReporters" : [ {
+        "taskKey" : "generatorConversion",
+        "taskValues" : { },
+        "subReporters" : [ ],
+        "reports" : [ {
+          "reportKey" : "disconnectedTerminalGenerator",
+          "values" : {
+            "generator" : {
+              "value" : "TestGenerator",
+              "type" : "UNTYPED"
+            }
+          }
+        } ]
+      }, {
+        "taskKey" : "batteryConversion",
+        "taskValues" : { },
+        "subReporters" : [ {
+          "taskKey" : "lineConversion",
+          "taskValues" : { },
+          "subReporters" : [ ],
+          "reports" : [ ]
+        }, {
+          "taskKey" : "tieLineConversion",
+          "taskValues" : { },
+          "subReporters" : [ ],
+          "reports" : [ ]
+        }, {
+          "taskKey" : "twoWindingsTransformerConversion",
+          "taskValues" : { },
+          "subReporters" : [ ],
+          "reports" : [ ]
+        }, {
+          "taskKey" : "threeWindingsTransformerConversion",
+          "taskValues" : { },
+          "subReporters" : [ ],
+          "reports" : [ ]
+        } ],
+        "reports" : [ ]
+      }, {
+        "taskKey" : "danglinglinesConversion",
+        "taskValues" : { },
+        "subReporters" : [ ],
+        "reports" : [ ]
+      }, {
+        "taskKey" : "courcirc",
+        "taskValues" : { },
+        "subReporters" : [ ],
+        "reports" : [ {
+          "reportKey" : "REC_0",
+          "values" : { }
+        }, {
+          "reportKey" : "REC_1",
+          "values" : { }
+        }, {
+          "reportKey" : "REC_2",
+          "values" : { }
+        }, {
+          "reportKey" : "REC_3",
+          "values" : { }
+        }, {
+          "reportKey" : "REC_4",
+          "values" : { }
+        }, {
+          "reportKey" : "REC_5",
+          "values" : { }
+        }, {
+          "reportKey" : "REC_6",
+          "values" : { }
+        }, {
+          "reportKey" : "REC_7",
+          "values" : { }
+        }, {
+          "reportKey" : "REC_8",
+          "values" : { }
+        }, {
+          "reportKey" : "TransientReactanceTooLow",
+          "values" : {
+            "nodes" : {
+              "value" : "T.NODE0, T.NODE1, T.NODE2, T.NODE3, T.NODE4",
+              "type" : "UNTYPED"
+            },
+            "reportSeverity" : {
+              "value" : "WARN",
+              "type" : "SEVERITY"
+            },
+            "nb" : {
+              "value" : 5,
+              "type" : "UNTYPED"
+            }
+          }
+        } ]
+      } ]
+    } ]
+  },
+  "dics": {
+    "default": {
+      "00000000-0000-0000-0000-000000000000@ShortCircuitAnalysis": "00000000-0000-0000-0000-000000000000@ShortCircuitAnalysis",
+      "ShortCircuitAnalysis": "ShortCircuitAnalysis (${providerToUse})",
+      "generatorConversion" : "Conversion of generators",
+      "disconnectedTerminalGenerator" : "Regulating terminal of connected generator ${generator} is disconnected. Regulation is disabled.",
+      "batteryConversion" : "Conversion of the batteries",
+      "lineConversion" : "Conversion of lines",
+      "tieLineConversion" : "Conversion of tie lines",
+      "twoWindingsTransformerConversion" : "Conversion of two windings transformers",
+      "threeWindingsTransformerConversion" : "Conversion of three windings transformers",
+      "danglinglinesConversion" : "Conversion of the dangling lines",
+      "courcirc" : "Logs generated by Courcirc simulator",
+      "REC_0" : "Taking into account resistances  : oui",
+      "REC_1" : "Taking into account capacities   : oui",
+      "REC_2" : "Taking into account VSC modelisation : non",
+      "REC_3" : "Taking into account current ratio of tap changers : oui",
+      "REC_4" : "Taking into account loads        : oui",
+      "REC_5" : "Taking into account shunts       : non",
+      "REC_6" : "Taking into account mutual lines : non",
+      "REC_7" : "Taking into account sub-transient values : non",
+      "REC_8" : "Interior area = whole network",
+      "TransientReactanceTooLow" : "${nb} node(s) with transient reactance too low ==> generator ignored\n${nodes}"
+    }
+  }
+}
